% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{get_initial_values}
\alias{get_initial_values}
\title{get_initial_values}
\usage{
get_initial_values(x, y)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}
}
\value{
A vector with four elements: (1) C term, (2) V term, (3) Local x average, and (4) Local y average
}
\description{
Helper function for supsmu_fit(): Computes the terms needed to estimate the first local regression line
}
\details{
See Friedman, 1984 for details.
}
\examples{
x=(1:1000)
y=x+rnorm(1000,5,500)
get_initial_values(x,y)


}
