% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{loess_lm_predict}
\alias{loess_lm_predict}
\title{loess_lm_predict}
\usage{
loess_lm_predict(x, y, degree, x_point)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}

\item{degree}{The degree of the local polynomial fit (1=linear fit, 2=quadratic fit)}

\item{x_point}{The x value that is used to predict a y value}
}
\value{
A scalar prediction of the y variable at x_point
}
\description{
Helper function for loess_fit(): Computes a prediction from the estimated local polynomial regression
}
\details{
The local polynomial model is fit by weight least squares with a tricubic weight function.
When degree=1, only an intercept and linear term for x are included in the model.
When degree=2, an intercept, linear term, and quadratic term for x are included in the model.
Predictions are made at x_point using the resulting estimates. See Cleveland, 1979 for details.
}
\examples{
x=rnorm(100)
y=x+rnorm(1000,5,500)
nearest_indices=nearest(x,0.75,x[1])
x_sub=x[nearest_indices]
y_sub=y[nearest_indices]
loess_lm_predict(x_sub,y_sub,2,x[1])

}
