% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{nearest}
\alias{nearest}
\title{nearest}
\usage{
nearest(x, span, x_point)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{span}{The proportion of datapoints to include in the local fit}

\item{x_point}{The x value that is used to predict a y value}
}
\value{
A vector of indices
}
\description{
Helper function for loess_fit(): Finds the indices of the span*N nearest neighbors to a given x value, where N is the total number of datapoints
}
\examples{
x=rnorm(100)
nearest(x,0.75,x[1])


}
