% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{supsmu_fit}
\alias{supsmu_fit}
\title{supsmu_fit}
\usage{
supsmu_fit(x, y, span = NULL, bass = 0)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}

\item{span}{The proportion of datapoints to include in the local fit (Optional)}

\item{bass}{A value between 0 and 10 that controls the smoothness of the predictions (See Details)}
}
\value{
A numeric vector of super smoother predictions for the y variable at each value of the x variable (sorted by the x variable)
}
\description{
Computes local linear predictions using Friedman's super smoother (Friedman, 1984)
}
\details{
If the span argument is specified, then a fixed span is used for each x value to
estimate the local linear regression line, and the bass argument is ignored. Otherwise,
the span is selected for each point via cross-validation, as in Friedman, 1984. For the variable span
algorithm, bass values closer to 10 result in a smoother-looking curve. Note that ties in the x variable
are treated as distinct points, which is different from the behavior of stats::supsmu()
(this function's method for handling ties is undocumented).
}
\examples{
x=(1:1000)
y=x+rnorm(1000,5,500)
supsmu_fit(x,y,span=0.2)
supsmu_fit(x,y,bass=5)
supsmu_fit(x,y)

}
