% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{supsmu_fit_span}
\alias{supsmu_fit_span}
\title{supsmu_fit_span}
\usage{
supsmu_fit_span(x, y, span = 0.2)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}

\item{span}{The proportion of datapoints to include in the local fit}
}
\value{
A numeric vector of super smoother predictions for the y variable at each value of the x variable (sorted by the x variable)
}
\description{
Helper function for supsmu_fit(): Fits a super smoother corresponding to the fixed span value
}
\details{
This function is the workhorse of supsmu_fit() if a fixed span is requested.
Otherwise, it serves as an intermediate helper function in the variable span algorithm (Friedman, 1984).
}
\examples{
x=(1:1000)
y=x+rnorm(1000,5,500)
supsmu_fit_span(x,y,span=0.2)


}
