% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{supsmu_fit_three_span}
\alias{supsmu_fit_three_span}
\title{supsmu_fit_three_span}
\usage{
supsmu_fit_three_span(x, y)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}
}
\value{
A list with three elements: (1) an Nx3 matrix of predictions for each span,
(2) an Nx3 matrix of local x averages for each span, and (3) an Nx3 matrix of V terms for each span (Friedman, 1984).
N is the total number of datapoints.
}
\description{
Helper function for supsmu_fit(): Fits super smoothers corresponding to spans of 0.5,0.2, and 0.05 with only one loop
}
