% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{supsmu_lm_predict}
\alias{supsmu_lm_predict}
\title{supsmu_lm_predict}
\usage{
supsmu_lm_predict(x_point, C, V, x_bar, y_bar)
}
\arguments{
\item{x_point}{The x value that is used to predict a y value}

\item{C}{The C term corresponding to the local regression estimation (Friedman, 1984)}

\item{V}{The V term corresponding to the local regression estimation (Friedman, 1984)}

\item{x_bar}{The local x average}

\item{y_bar}{The local y average}
}
\value{
A scalar prediction of the y variable at x_point
}
\description{
Helper function for supsmu_fit(): Computes a prediction from the estimated local linear regression
}
\examples{
x=(1:1000)
y=x+rnorm(1000,5,500)
u=get_inital_values(x,y)
supsmu_lm_predict(x[1],u[1],u[2],u[3],u[4])


}
