% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{update_lm_predict}
\alias{update_lm_predict}
\title{update_lm_predict}
\usage{
update_lm_predict(x, y, J, C, V, x_bar, y_bar, index)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}

\item{J}{The span value times the total number of datapoints}

\item{C}{The C term used for regression estimation (Friedman, 1984)}

\item{V}{The V term used for regression estimation (Friedman, 1984)}

\item{index}{The index of the current prediction point}
}
\value{
A vector with four elements: (1) C term, (2) V term, (3) Local x average, and (4) Local y average
}
\description{
Helper function for supsmu_fit(): Updates the estimates of the previous local linear regression without refitting the model
}
\details{
See Friedman, 1984 for details.
}
\examples{
x=(1:1000)
y=x+rnorm(1000,5,500)
u=get_inital_values(x,y)
u=update_lm_predict(x,y,u[1],u[2],u[3],u[4],2)

}
