% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{loess_fit}
\alias{loess_fit}
\title{loess_fit}
\usage{
loess_fit(x, y, degree = 2, span = 0.75)
}
\arguments{
\item{x}{A numeric vector corresponding to the independent variable}

\item{y}{A numeric vector corresponding to the dependent variable}

\item{degree}{The degree of the local polynomial fit (1=linear fit, 2=quadratic fit)}

\item{span}{The proportion of datapoints to include in the local fit}
}
\value{
A numeric vector of LOESS predictions for the y variable at each value of the x variable
}
\description{
Computes LOESS local polynomial predictions (Cleveland, 1979)
}
\examples{
set.seed(1)
x=seq(0,10,length.out=100)
y=sin(x)+rnorm(100,0,0.5)
loess_fit(x,y,span=0.2,degree=2)
loess_fit(x,y,span=0.5,degree=1)

}
